Splitted eCompiler to separated files
